pyproject.toml
[project]
name = "osr-protocol"
version = "0.1.0"
description = "Solana blockchain economic layer for System R AI"
requires-python = ">=3.11"
dependencies = [
    # Core
    "python-dotenv>=1.0.0",
    "structlog>=24.0.0",
    "httpx>=0.27.0",
    # LLM providers
    "anthropic>=0.40.0",
    "openai>=1.50.0",
    "google-generativeai>=0.8.0",
    # AWS
    "boto3>=1.35.0",
    # Solana
    "solders>=0.21.0",
    "solana>=0.34.0",
    # Social
    "telethon>=1.37.0",
    "tweepy>=4.14.0",
    # Data
    "pydantic>=2.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.7.0",
    "moto[dynamodb]>=5.0.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
pythonpath = ["."]
testpaths = ["tests"]

[tool.ruff]
target-version = "py311"
line-length = 140

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
